conversion types possible loss of data warning resolved by static cast

+++ Project/Source/Modules/ModuleFiles.cpp
@@ -41,7 +41,7 @@ Buffer<char> ModuleFiles::Load(const char* filePath) const {
 		return buffer;
 	}
 
-	buffer.Allocate(size + 1);
+	buffer.Allocate(static_cast<size_t>(size) + 1);
 	char* data = buffer.Data();
 	PHYSFS_sint64 numBytes = PHYSFS_readBytes(file, data, size);
 	if (numBytes < size) {
